# Day4/ProductsAPIApp/ProductsAPIApp/Models/Product.cs
namespace ProductsAPIApp.Models
{
    public class Product
    {
        public int ProductID { get; set; }
        public string ProductName { get; set; }
        public int Quantity { get; set; }
    }
}