# Day5/MVCSQLApp/MVCSQLApp/Models/Product.cs
namespace MVCSQLApp.Models
{
    public class Product
    {
        public int ProductID { get; set; }
        public string ProductName { get; set; }    
        public int Quantity { get; set; }
    }
}